Improved checking of ZX82 ROM containers
Sanity check ROM images in ZX82 files (only) to ensure they start with a DI
instruction. We now correctly report the ROM11 image as invalid.

All but one of Andy Wright's ROM images use the same bootstrap code, with
the ROM starting from the file same offset (140 bytes). The ROM11 image has
the same ZX82 header but different bootstrap code, and the ROM starting at
a different offset. The embedded ROM code is inconsistent with the ROM1
code not matching calls from ROM0, so it appears to be a bad image.

Added "rom*" to the default ROM browse filter so we can see Dr Wright's
files without changing to the All Files filter.