Added missing tape auto-load triggers
Selecting a tape from the MRU menu and dropping tape files on the tape
browser were not triggering as expected. Tape auto-loading is now handled
similar to it is for disks.

Removed startup screen skipping from auto-load to give a single KeyIn
location for doing it, which makes more sense.