Added support for older SAM ROM images (#77)
Rather than relying on fixed addresses (from ROM v3.0) we now search for the
code of interest. This allows us to support older ROM versions that have
similar code in different locations, or may be missing the code entirely.

Since we know whether we've found the hook point we can gracefully fall back
to not supporting some emulator features. ROM 1.0 is missing BOOTEX, which
is needed to support booting of non-bootable disks. Without it we behave
like the option is disabled and allow the No DOS error to be seen.