Added basic motion blur support to SDL backend
When the motion blur option is enabled we'll now blend a portion of the
previous display image with the current display image. However it's only
simple blending so moving pixels blended against black in the previous frame
will have reduced intensity.

The D3D11 and OpenGL backends use a custom shader to retain full brightness
but that's not widely supported through SDL. It does support an equivalent
maximum operation but only under D3D, and even that didn't work for me.