Support portable Windows package building
Build x86/x64 binaries compatible with Windows 7 SP1 or later. This requires
a little more care than the default build output.

We also build an Inno Setup installer containg both x86 and x64 versions,
with the appropriate version installed. The demo disk images are included if
found locally, but they won't be for server-side builds.

Modules built on GitHub are NOT code signed, but official releases will
continue to be signed for a slightly smoother end-user experience.