Debugger assistance with DI;HALT state
Entering the debugger after DI;HALT doesn't make it clear the CPU is halted
and there's no easy way to clear the halted state. Worse, changing PC keeps
the CPU halted so nothing more is executed.

To smooth this out we'll automatically clear the halted state when stepping
or if PC has changed during the debugger session. This only happens if
interrupts are also disabled.

The DI;HALT state now shows the DI indicator in bright red as a hint, and
the disassembly starts an instruction early to reveal the HALT. The halted
state will be better exposed in the future.