Fixed consecutive FDC data timeouts not working
We need to reset the timeout counter when it triggers to allow further
timeouts to occur. Previously we only reset it on successful data reads.
This worked before 50ed3ab1 as we used a mod expression to allow repeats.

Legend of Eshan uses two consecutive READ_ADDRESS commands that end in
a timeout, and the second one was not triggering correctly. The game timed
out instead and reported a No Disk error.