Fixed incompatible MGT formatting (#86)
Attempting to format an MGT disk image with a format not compatible with its
10-sector fixed format now generates a write-protected error. Previously we
tried to indicate a write fault, which doesn't seem to be valid FDC
behaviour.

This caused problems in Pro-Dos when attempting to reformat the boot disk
with the 9-sector CP/M format. The invalid status was not expected and the
formatting appeared to complete successfully, though nothing was actually
written.