Fixed no-disk formatting behaviour (#86)
Changed write-track FDC command to wait for a spinning disk to be detected
before starting. The FDC remains in a busy state during this time. When a
disk is inserted the motor starts briefly to engage the disk and the index
signal will be seen, allowing the command to continue executing.

Pro-Dos doesn't check for a disk in the drive and relies on this behaviour
for its disk formatting. Without it formatting appears to succeed even with
no disk in the drive.