Fixed Win32 device list issue with bootable disks
We previously relied on raw disks having a dummy partition table entry spanning
the entire disk. This was a workaround for earlier versions of Windows not
returning an accurate sector count in the geometry. However, bootable HDDs have
a real partition table that replaces this dummy entry with four MBR partitions.
This failed our extra checks and caused the disk to be skipped.

We now just fetch the normal disk geometry instead, and also filter out non-USB
devices to hide system drives.