Fix debugger graphics view mode 1+2 crash (#94)
The graphics view columns have limited width, depending on the viewing mode.
This limit was enforced when changing modes within the view, but not when first
entering the view, where the view mode was set to match the live mode.

If the previous view mode was 3 or 4 with a width > 32, and the debugger was
entered when the active mode was 1 or 2, the previous width would be invalid
and cause a crash on the next draw.

We now only set the view mode to match the live mode on first entry, when it's
safe. The previous mode is remembered between uses of the graphics view in the
same session.