Fixed disassembly addr mismatch from previous view (#96)
The disassembly view avoids resetting the first address in the view if the new
location is already covered by the cached view. This caused confusion when
switching from other views in nearby locations.

We now force the new view to the same starting location as the previous view
when changing views.