Disable motion blur function for now
Motion blur looks awful without a VRR display to deliver frames at the correct
rate. For now we'll hide the option and deactivate the function. If a user
really wants it they can change allowmotionblur option in SimCoupe.cfg.

Also changed the blend percentage to 50%, so half of each frame is blended.
This gives a stronger effect than before but should be more accurate.