Add option to randomise IM2 bus value
Previously the bus value was always 0xff, but peripherals can affect this. To
help track down IM2 issues the bus value can now be randomised. It cycles
through all possible values in steps of 37 (coprime with 256) to test them all.

Updated the Z80 core to provide the mixin override to control the bus value.