Fixed stuck fast boot display after soft reset
The timeout counter was only assigned on hard resets so soft resets would show
the static boot screen if the ROM failed to reach the copyright message. This
could happen with custom boot ROMs that are chaining to incompatible boot
sector code, usually due to a mismatched HDBOOT+BS combination.

We now correctly re-enable the display anyway, which may still be blank.