Restore support for Atom Lite in left drive bay
For now this remains a hidden feature that requires editing the configuration
file to enable. If it proves useful it will be added back to the options UI.

To use it set drive1=3 to set the drive 1 type to Atom Lite. Then set disk
paths in atomdiskleft0 and atomdiskleft1, similar to atomdisk0 and atomdisk1.