Added option to override resource file location (#101)
The default location for resources is platform-specific, with the EXE directory
used on Windows, and <root>/share/simcoupe on Linux.

This can now be overridden using the -respath <path> option, to specify a
different location for resources files such as ROM and symbol files. The path
specified is saved in the options and used for later launches.